pyproject.toml
[project]
name = "f1tenth-racetrack-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pyyaml>=6.0.3",
    "scipy>=1.17.1",
    "shapely>=2.1.2",
]
